refactor: transfer completeController function to a sepaat modulere
